+++ services/sso/src/__tests__/dashboard/users/list.test.ts
@@ -4,7 +4,7 @@ import { z } from 'zod'
 import { pathRoot } from '../../../routes/routes'
 import { userSchema } from '../../../schemas'
 import { itinerariesData, server, testUserData } from '../../globalSetup'
-import { client } from '../../../db/client'
+import db from '../../../db/knexClient'
 import { DashboardUsersList } from '../../../schemas/users/dashboardUsersListSchema'
 import { UserRole, UserStatus } from '../../../schemas/users/userSchema'
 import { dashboardLoginAndGetToken } from '../../helpers/testHelpers'
@@ -27,6 +27,7 @@ const testName = 'testing'
 const testDni = '38826335N'
 const testRole = UserRole.REGISTERED
 const testStatus = UserStatus.ACTIVE
+
 beforeAll(async () => {
   authAdminToken = await dashboardLoginAndGetToken(
     testUserData.admin.dni,
@@ -36,31 +37,29 @@ beforeAll(async () => {
     testUserData.mentor.dni,
     testUserData.mentor.password
   )
-  const queryResult = await client.query(
-    `SELECT
-    u.id,
-    u.name AS name,
-    u.dni AS dni,
-    u.status,
-    u.role,
-    u.deleted_at AS "deletedAt",
-    TO_CHAR(u.created_at, 'YYYY-MM-DD') AS "createdAt",
-    i.name AS "itineraryName"
-  FROM
-    "user" u
-  JOIN itinerary i ON u.itinerary_id = i.id;`
-  )
 
-  users = queryResult.rows
+  users = await db('user as u')
+    .join('itinerary as i', 'u.itinerary_id', '=', 'i.id')
+    .select(
+      'u.id',
+      'u.name as name',
+      'u.dni as dni',
+      'u.status',
+      'u.role',
+      'u.deleted_at as deletedAt',
+      db.raw('TO_CHAR(u.created_at, \'YYYY-MM-DD\') as "createdAt"'),
+      'i.name as itineraryName'
+    )
 })
+
 describe('Testing get users endpoint', () => {
-  it('returns a  collection of users successfully with a logged-in admin user', async () => {
+  it('returns a collection of users successfully, with a logged-in admin user', async () => {
     const response = await supertest(server)
       .get(route)
       .set('Cookie', [authAdminToken])
     expect(response.status).toBe(200)
-    expect(response.body).toHaveLength(users.length)
     expect(response.body).toEqual(users)
+    expect(response.body).toHaveLength(users.length)
     expect(responseSchema.safeParse(response.body).success).toBeTruthy()
   })
   it('Returns a collection of users successfully with a logged-in mentor user, where the collection must contain users with the same itinerary.', async () => {
@@ -77,7 +76,7 @@ describe('Testing get users endpoint', () => {
       .every((user) => user.itineraryName === firstItineraryName)
     expect(allSameItinerary).toBe(true)
   })
-  it('returns a  collection of users by itinerary slug successfully with a logged-in admin user ', async () => {
+  it('returns a collection of users by itinerary slug successfully with a logged-in admin user ', async () => {
     const response = await supertest(server)
       .get(route)
       .query({ itinerarySlug: itinerariesData[3].slug })
@@ -99,7 +98,7 @@ describe('Testing get users endpoint', () => {
     expect(response.body).toEqual([])
     expect(responseSchema.safeParse(response.body).success).toBeTruthy()
   })
-  it('returns a  collection of users by status successfully with a logged-in admin user ', async () => {
+  it('returns a collection of users by status successfully with a logged-in admin user ', async () => {
     const activeUsers = users.filter((u) => u.status === UserStatus.ACTIVE)
     const response = await supertest(server)
       .get(route)
@@ -116,7 +115,7 @@ describe('Testing get users endpoint', () => {
     expect(sortedResponseBody).toEqual(sortedExpected)
     expect(responseSchema.safeParse(body).success).toBeTruthy()
   })
-  it('returns a  collection of users within a date range successfully with a logged-in admin user, [NOT WORK WITHIN 00:00-02:00 AM!!]', async () => {
+  it('returns a collection of users within a date range successfully with a logged-in admin user, [NOT WORK WITHIN 00:00-02:00 AM!!]', async () => {
     const dateMinusOne = new Date()
     dateMinusOne.setDate(dateMinusOne.getDate() - 1)
     const datePlusOne = new Date()
@@ -193,7 +192,7 @@ describe('Testing get users endpoint', () => {
     expect(body).toHaveLength(1)
     expect(responseSchema.safeParse(body).success).toBeTruthy()
   })
-  it('returns a  collection of users by role successfully with a logged-in admin user ', async () => {
+  it('returns a collection of users by role successfully with a logged-in admin user ', async () => {
     const registeredUsers = users.filter((u) => u.role === UserRole.REGISTERED)
     const response = await supertest(server)
       .get(route)
@@ -212,7 +211,7 @@ describe('Testing get users endpoint', () => {
     expect(sortedResponseBody).toEqual(sortedExpected)
     expect(responseSchema.safeParse(body).success).toBeTruthy()
   })
-  it('returns a colletion of users by 4 diferent search values', async () => {
+  it('returns a user by 4 diferent search values', async () => {
     const response = await supertest(server)
       .get(route)
       .query({
@@ -225,24 +224,20 @@ describe('Testing get users endpoint', () => {
     const { body }: { body: DashboardUsersList } = response
     expect(response.status).toBe(200)
     expect(body).toBeInstanceOf(Array)
-    expect(body).toHaveLength(2)
+    expect(body).toHaveLength(1)
     expect(responseSchema.safeParse(body).success).toBeTruthy()
   })
 })
 describe('User Status Handling in Get Users Endpoint', () => {
   afterEach(async () => {
-    await client.query('UPDATE "user" SET status = $1 WHERE dni = $2', [
-      UserStatus.ACTIVE,
-      testUserData.admin.dni,
-    ])
+    await db('user')
+      .update({ status: UserStatus.ACTIVE })
+      .where('dni', testUserData.admin.dni)
   })
   it('should fail to return a collection of users with a blocked logged-in admin', async () => {
     const adminDni = testUserData.admin.dni
     const newStatus = UserStatus.BLOCKED
-    await client.query('UPDATE "user" SET status = $1 WHERE dni = $2', [
-      newStatus,
-      adminDni,
-    ])
+    await db('user').update({ status: newStatus }).where('dni', adminDni)
     const response = await supertest(server)
       .get(route)
       .set('Cookie', [authAdminToken])
@@ -260,10 +255,7 @@ describe('User Status Handling in Get Users Endpoint', () => {
 
     const adminDni = testUserData.admin.dni
     const newStatus = UserStatus.PENDING
-    await client.query('UPDATE "user" SET status = $1 WHERE dni = $2', [
-      newStatus,
-      adminDni,
-    ])
+    await db('user').update({ status: newStatus }).where('dni', adminDni)
     const response2 = await supertest(server)
       .get(route)
       .set('Cookie', [authAdminToken])