+++ services/sso/src/controllers/dashboard/users/list.ts
@@ -1,7 +1,7 @@
 import { Context, Middleware } from 'koa'
-import { client } from '../../../db/client'
-import { queryBuilder } from '../../../utils/queryBuilder'
 import { User, UserRole } from '../../../schemas'
+import db from '../../../db/knexClient'
+import { knexQueryBuilder } from '../../../utils/knex.queryBuilder'
 
 export const dashboardListUsers: Middleware = async (ctx: Context) => {
   const { role, id } = ctx.state.user as Pick<User, 'id' | 'role'>
@@ -9,16 +9,18 @@ export const dashboardListUsers: Middleware = async (ctx: Context) => {
   if (role === UserRole.MENTOR) {
     mentorUserId = id
   }
-  const { query, queryParams } = queryBuilder(ctx.state.query, mentorUserId)
 
-  const queryResult = await client.query(query, queryParams)
+  const { query } = knexQueryBuilder(ctx, db, mentorUserId)
 
-  if (!queryResult.rowCount) {
+  const queryResult = await query
+
+  if (!queryResult.length) {
     ctx.status = 200
     ctx.body = []
     return
   }
-  const usersName = queryResult.rows
+
+  const usersList = queryResult
   ctx.status = 200
-  ctx.body = usersName
+  ctx.body = usersList
 }